Added tool that builds kinship input files.
The build_kinship_input.py script generates input files for the MR Kinship software:
* a file with the genotypes
* a file with the list of the individuals
each column of the first file represents an individual (if the --transpose_output parameter has been used, each individual is represented by a row of the first file).
The Nth row of the second file matches the Nth column (or row) of the first file.