initial support for yaml based biobank objects serialization
This commit implements a simple implementation of a deserializer capable
of converting yaml based description of biobanks objects to actual
objects. It supports referencing to objects already existing in
the biobank and referencing to objects described in the same yaml
stream.

There is currently no documentation.
See test/serialize/test_deserialize.py for usage examples.