Precleaning support in objects deletion.
KB assures that an object on which other objects depend cannot be
cancelled. On the other hand, there are object of special types, e.g.,
DataObject, on which other objects cannot depend and should be deleted
when the master object, e.g., DataSample, is deleted.

To simplify handling of the latter case, I have added
proxy_core.__precleanup__(),  a method that is invoked, before
anything else, by kb.delete(obj).