Ipython notebook based examples
Added basic infrastructure for example writing based on ipython
notebooks. The idea is to write the biobank examples directly as
notebooks and to generate 'static' documentation by converting the
notebooks into rst. It is expected that this will simplify the tedious
task of writing example code and maintaining it in phase with
development.

The makefile notebooks/Makefile has a ipynb to rst conversion
as its default target, while  'make startnb' will start a, local,
ipython notebook server.
Ipython should be version 1.1 or above.