Fix galaxy wait to handle Galaxy's new logic wrt dataset states
Newer versions of Galaxy given a history state in 'queued' even if
individual datasets are in an error state.  This commit modifies the
wait method by importing logic from bioblend.objects to examine the
individual dataset states and exit when any are in error or all are out
of the "pending" states.